[Refactor] Only prepare sandbox for cc backend.
Now when you import taichi, we'll only show:
```
>>> import taichi
[Taichi] version 0.7.29, llvm 10.0.0, commit e719c60a, linux, python 3.8.11
```

We'll prepare a sandbox when you initialize only on **cc backend**:
```
>>> ti.init(arch=ti.cc)
[Taichi] preparing sandbox at /tmp/taichi-osr4s9x5
[Taichi] Starting on arch=cc
```
