+++ python/taichi/lang/__init__.py
@@ -2,6 +2,7 @@ import functools
 import os
 from copy import deepcopy as _deepcopy
 
+from taichi.core.util import locale_encode
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang import impl
 from taichi.lang.exception import InvalidOperationError
@@ -247,6 +248,21 @@ class _SpecialConfig:
         self.experimental_real_function = False
 
 
+def prepare_sandbox():
+    '''
+    Returns a temporary directory, which will be automatically deleted on exit.
+    It may contain the taichi_core shared object or some misc. files.
+    '''
+    import atexit
+    import shutil
+    from tempfile import mkdtemp
+    tmp_dir = mkdtemp(prefix='taichi-')
+    atexit.register(shutil.rmtree, tmp_dir)
+    print(f'[Taichi] preparing sandbox at {tmp_dir}')
+    os.mkdir(os.path.join(tmp_dir, 'runtime/'))
+    return tmp_dir
+
+
 def init(arch=None,
          default_fp=None,
          default_ip=None,
@@ -339,6 +355,8 @@ def init(arch=None,
         ti.info(f'Following TI_ARCH setting up for arch={env_arch}')
         arch = _ti_core.arch_from_name(env_arch)
     ti.cfg.arch = adaptive_arch_select(arch)
+    if ti.cfg.arch == cc:
+        _ti_core.set_tmp_dir(locale_encode(prepare_sandbox()))
     print(f'[Taichi] Starting on arch={_ti_core.arch_name(ti.cfg.arch)}')
 
     if _test_mode:
